完成 用户注册/登陆 完成 前端首页店铺列表展示, 完成 前端店铺下分类 以及菜品展示 完成 用户收货地址的添加,修改 完成 用户购物车(收货地址+购物车商品展示) ↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓  #todo 订单的生成//展示

+++ apps/model/food_model.py
@@ -1,5 +1,5 @@
 from apps.model import db
-from .user_model import BaseModel
+from .seller_model import BaseModel
 
 
 class MenuCategory(BaseModel):
@@ -39,7 +39,7 @@ class MenuFood(BaseModel):
     category_pid = db.Column(db.Integer, db.ForeignKey('menu_category.pub_id'))
     cates = db.relationship('MenuCategory', backref='foods')  # 添加一条关系
     # 菜品价格
-    goods_price = db.Column(db.DECIMAL(6, 2), default=0.0)
+    goods_price = db.Column(db.Float, default=0.0)
     # 菜品描述
     description = db.Column(db.String(128), default='')
     # 月销售额
